PROD-190 local actor 파일명 정리
