+++ packages/fedify/src/federation.ts
@@ -1,7 +1,7 @@
 import { createFederation, MemoryKvStore } from '@fedify/fedify';
 import { resolveConfiguredLocalInstance } from '@kosmo/core/local-instance';
 import { ensureDrizzleLocalProfileActor } from './local-actor-store';
-import { createLocalProfilePerson } from './local-profile-actor';
+import { createLocalProfilePerson } from './local-profile-person';
 import type { Federation } from '@fedify/fedify';
 
 const federationOrigin = process.env.PUBLIC_ORIGIN;
